factor out utils
